// web-app/src/app/models/statut.enum.ts
export enum StatutEnum {
    VIDE_CHEZ_LE_MARKETEUR = "vide chez marketeur",
    PRETE_A_ETRE_REMPLIE = "prete a etre remplie",
    PRETE_A_LIVREE_AU_MARKETEUR = "prete a etre livree au marketeur",
    EN_COURS_DE_LIVRAISON_AU_MARKETEUR = "en cours de livraison au marketeur",
    CHEZ_LE_MARKETEUR = "chez marketeur",
    EN_COURS_DE_LIVRAISON_AU_VENDEUR = "en cours de livraison au revendeur",
    PLEINE_CHEZ_LE_VENDEUR = "pleine chez le revendeur",
    PLEINE_CHEZ_LE_CLIENT = "pleine chez le client",
    VIDE_CHEZ_LE_CLIENT = "vide chez le client",
    VIDE_CHEZ_LE_REVENDEUR = "vide chez le revendeur",
    VIDE_EN_COURS_DE_LIVRAISON_AU_MARKETEUR = "vide en cours de livraison au marketeur"
}